backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prompt-platform-backend"
version = "0.1.0"
description = "Multi-Tenant AI-Powered RAG-Based Document Intelligence Platform"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",

    # Pydantic & config
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",

    # Database
    "supabase>=2.4.0",
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.29.0",

    # Vector store
    "chromadb>=0.4.22",

    # Embeddings
    "sentence-transformers>=2.6.0",
    "torch>=2.2.0",

    # AI / LLM
    "google-generativeai>=0.5.0",
    "langchain>=0.1.12",
    "langgraph>=0.0.40",
    "langchain-google-genai>=1.0.1",
    "openai>=1.14.0",

    # OCR
    "pytesseract>=0.3.10",
    "easyocr>=1.7.1",
    "Pillow>=10.2.0",
    "pdf2image>=1.17.0",
    "python-bidi>=0.4.2",
    "arabic-reshaper>=3.0.0",

    # Document generation
    "reportlab>=4.1.0",
    "python-docx>=1.1.0",

    # Auth
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "twilio>=8.13.0",

    # Text & Language processing
    "langdetect>=1.0.9",

    # HTTP & Async
    "httpx>=0.27.0",
    "aiofiles>=23.2.1",

    # Utilities
    "python-dotenv>=1.0.1",
    "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
